Fix confusion about zend_string and char* in stmt_describer
PDO takes a zend_string, but there were (rare) code paths where it
could be given char* instead. Move the zend_string creation to after
the if/else block (so it unconditionally converts) and free it after
creating the PDO structure to avoid leaks. Some opportunities for
caching were identified, but left unexploited for now.

This doesn't affect PHP 5, but PHP 5 is going bye-bye anyways later.

Another issue where stmt_describer accidentally reused the old
buffer size in the case where the buffer was too small (and a new
buffer allocated) was fixed. This matches the logic in ibm_db2.