PDO returns a ValueError on -1 column index
In PHP 7, the driver would handle this error and let it cause an
SQL error. Change the test to consider a ValueError acceptable,
since PDO won't mutate the SQL state.

For the IBM i version of the test, it wasn't regex based, so just
copy the regex from the LUW version of the test.