Fix memory leak with interned strings and out parameters
Tony added code to deal with interned strings with out parameters.
The reason why is because interned strings are immutable and PDO_IBM
will bind directly into zvals, so we need a mutable copy of the string.
He forgot to actually free the mutable string he created though. This
holds onto the string in the parameter driver area so we can free it
at the event hook for it. Hacky, but works.

I initially had a more complex fix that combined the the init/extend
by just moving the pad length check before and using that if needed,
but I was unsure of the impact (i.e null term) so I had backed off
for this more modest patch instead.

The /real/ fix for this should be converting PDO_IBM into using
a temporary buffer and copying at the end instead of binding direct
into zvals, but this puts a bandage on the wound for now.