Enable liveness checks on PASE
Things like operators ending the job are an unfortunate reality on
IBM i, even if the connections aren't necessarily remote. ibm_db2
does this too, albeit as an optional thing. (Procedural ODBC does as
well, but by default.)

Later, we should be handling the new PDO structure changes introduced
in PHP 8.1, but that can be done at a later date.