Defer to setAttribute instead
This reduces a lot of redundant code for setting attributes. While we
can do it in connect, the set attribute handler operates on handles and
doesn't care if it's before connect, which works just as well for the
early-set cases like commitment control or system naming.

We do have to handle the use trusted context part anyways, as that must
be set before connect and seemingly can't be afterwards.

Still some refactoring opportunities to simply the iterator and use
pdo_attr_(str|l)val for anything we can't defer to the set attribute
handler for. It also makes me want to implement some more get attribute
to easily verify the current state.