Replace the LUW version of the test with IBM i one
I'm not sure how the original version of the test ever made sense when I
was reading it. The behaviour it expects seemed counter-intuitive with
how it was expecting casting to happen.

Bizarrely, the IBM i version does match the behaviour of modern PHP and Db2.
So, let's just put that in its place.