Implement dbcs_alloc for PASE in PDO_IBM
ibm_db2 has it, and it's an unfortunate workaround that's sometimes
needed to do how unreliable it can be to detected truncation. Ideally
this option should be made obsolete, but until it can be, this can be
enabled for situations where EBCDIC characters expand into longer
UTF-8 ones.

Running the tests, none seem affected, same amount of pass/fail.