Workaround: Disable the skip cache on PHP 8.1
On PHP 8.1, the skip cache seemingly collates all the reused skip
sections into a single file. However, the skip check relies on using
the name of the test to determine if it should be skipped based on
platform. This is why the skip blocks that *aren't* the same will
trigger reliably.

This will slow it down a smidge, but work reliably until the skip
checks are refactored.